requirements.txt
pycryptodome>=3.10.1
pytest>=7.0.0  # 如果需要测试